waiting for limit increase for ml.p2.xlarge
Got an error ResourceLimitExceeded: An error occurred (ResourceLimitExceeded) when calling the CreateTrainingJob operation: The account-level service limit 'ml.p2.xlarge for training job usage' is 0 Instances, with current utilization of 0 Instances and a request delta of 1 Instances. Please contact AWS support to request an increase for this limit.

Followed instructions on course page to request an limit increase to ml.p2.xlarge it should work with ml.m4.xlarge. I just want to see want is the difference.